Database in unique module.
